poemgr: Use floating point for power measurements
PoE standards specify power granularity to 0.1W. That can be clumsy to
represent with integers. Instead, use floating point and JSON doubles.

Note that json_object_new_double() will add a lot of pointless digits
to the output. This is resolved with json_object_new_double_s. The
format is kept at one digit after the decimal. This is sufficient for
power measurements.

+++ poemgr.h
@@ -47,8 +47,8 @@ struct poemgr_port_settings {
 struct poemgr_port_status {
 	int enabled;
 	int active;
-	int power_limit;
-	int power;
+	float power_limit;
+	float power;
 	int poe_class;
 
 	int faults;
@@ -68,7 +68,7 @@ struct poemgr_input_status {
 };
 
 struct poemgr_output_status {
-	int power_budget;
+	float power_budget;
 
 	time_t last_update;
 };
